Fix progress bar click event
Correction to ensure that correct X value is used to set fill width
when video is less than the full width of the screen. Also removed div
that held progress bar. Div was originally included to make the
clickable area for the progress bar wider, but I was unable to find a
method to accurately set the width of the div to exactly equal the
width of the canvas element.